+++ rich/tabulate.py
@@ -13,7 +13,7 @@ def tabulate_mapping(mapping: Mapping, title: str = None) -> Table:
     
     Args:
         mapping (Mapping): A mapping object (e.g. a dict);
-        title (str, optional): [description]. Optional title to be displayed over the table.
+        title (str, optional): Optional title to be displayed over the table.
     
     Returns:
         Table: A table instance which may be rendered by the Console.